--- app/src/main/java/com/example/basearch/presentation/di/PresentationModule.kt
@@ -1,5 +1,0 @@
-package com.example.basearch.presentation.di
-
-val presentationModule = listOf(
-    viewModelModule
-)
