--- app/src/main/java/com/example/basearch/presentation/extensions/ViewExtensions.kt
@@ -1,11 +1,0 @@
-package com.example.basearch.presentation.extensions
-
-import android.view.View
-
-fun View.setVisible() {
-    visibility = View.VISIBLE
-}
-
-fun View.setGone() {
-    visibility = View.GONE
-}
