--- app/src/main/java/com/example/basearch/presentation/ui/main/adapter/ReposAdapterListener.kt
@@ -1,7 +1,0 @@
-package com.example.basearch.presentation.ui.main.adapter
-
-import br.com.pebmed.domain.entities.RepoModel
-
-interface ReposAdapterListener {
-    fun onItemClick(repo: RepoModel)
-}
