--- app/src/main/java/com/example/basearch/presentation/ui/pullRequest/list/adapter/PullRequestListAdapterListener.kt
@@ -1,5 +1,0 @@
-package com.example.basearch.presentation.ui.pullRequest.list.adapter
-
-interface PullRequestListAdapterListener {
-    fun onItemClick(pullRequestNumber: Long)
-}
