+++ app/src/main/java/com/pebmed/basearch/presentation/di/PresentationModule.kt
@@ -1,0 +1,5 @@
+package com.pebmed.basearch.presentation.di
+
+val presentationModule = listOf(
+    viewModelModule
+)
