+++ app/src/main/java/com/pebmed/basearch/presentation/extensions/ViewExtensions.kt
@@ -1,0 +1,11 @@
+package com.pebmed.basearch.presentation.extensions
+
+import android.view.View
+
+fun View.setVisible() {
+    visibility = View.VISIBLE
+}
+
+fun View.setGone() {
+    visibility = View.GONE
+}
