+++ app/src/main/java/com/pebmed/basearch/presentation/ui/main/adapter/ReposAdapterListener.kt
@@ -1,0 +1,7 @@
+package com.pebmed.basearch.presentation.ui.main.adapter
+
+import br.com.pebmed.domain.entities.RepoModel
+
+interface ReposAdapterListener {
+    fun onItemClick(repo: RepoModel)
+}
