+++ app/src/main/java/com/pebmed/basearch/presentation/ui/pullRequest/list/adapter/PullRequestListAdapterListener.kt
@@ -1,0 +1,5 @@
+package com.pebmed.basearch.presentation.ui.pullRequest.list.adapter
+
+interface PullRequestListAdapterListener {
+    fun onItemClick(pullRequestNumber: Long)
+}
